requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.2.1
filelock==3.19.1
fsspec==2025.9.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
llvmlite==0.45.0
MarkupSafe==3.0.3
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.4.2
numba==0.62.1
numpy==2.2.6
openai-whisper==20250625
packaging==25.0
pillow==11.3.0
proglog==0.1.12
pydub==0.25.1
python-dotenv==1.1.1
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.1
regex==2025.9.18
requests==2.32.5
setuptools==80.9.0
sympy==1.14.0
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==1.26.16
